feat: add permission restriction on certain methods

+++ identity_service/src/identity_service_impl.py
@@ -3,26 +3,23 @@ from typing import Tuple
 from uuid import uuid4
 
 import grpc
-
 import prisma.errors
 
-from errors.InvalidTokenError import InvalidTokenError
-from errors.unique_error import UniqueError
-from errors.value_not_found_error import ValueNotFoundError
-from src.models.user import User
-from utils.encoder import Encoder
-from utils.jwt_controller import JwtController, TokenType
-
-from generated.identity_service_pb2_grpc import IdentityServiceServicer as GrpcServicer
-from google.protobuf.empty_pb2 import Empty
-from repository.token_repository_interface import TokenRepositoryInterface
-from repository.user_repository_interface import UserRepositoryInterface
 import generated.auth_pb2 as auth_proto
 import generated.delete_user_pb2 as delete_user_proto
 import generated.get_access_token_pb2 as get_access_token_proto
 import generated.get_user_pb2 as get_user_proto
 import generated.identity_service_pb2 as requests_proto
 import generated.update_user_pb2 as update_user_proto
+from errors.InvalidTokenError import InvalidTokenError
+from errors.unique_error import UniqueError
+from errors.value_not_found_error import ValueNotFoundError
+from generated.identity_service_pb2_grpc import IdentityServiceServicer as GrpcServicer
+from repository.token_repository_interface import TokenRepositoryInterface
+from repository.user_repository_interface import UserRepositoryInterface
+from src.models.user import User, UserType
+from utils.encoder import Encoder
+from utils.jwt_controller import JwtController, TokenType
 
 
 class IdentityServiceImpl(GrpcServicer):
@@ -242,7 +239,9 @@ class IdentityServiceImpl(GrpcServicer):
             user_id, session_id = self._jwt_controller.decode(
                 request.refresh_token, TokenType.ACCESS_TOKEN
             )
-            _ = await self._user_repository.get_user_by_session_id(session_id=session_id)
+            _ = await self._user_repository.get_user_by_session_id(
+                session_id=session_id
+            )
             access_token = self._jwt_controller.generate_access_token(
                 user_id=user_id, session_id=session_id
             )
@@ -341,23 +340,33 @@ class IdentityServiceImpl(GrpcServicer):
                 status_code=500, message="Internal server error"
             )
 
-    async def get_all_users(self, _: Empty, context: grpc.ServicerContext) -> get_user_proto.UsersResponse:
+    async def get_all_users(
+        self, request: get_user_proto.GetAllUsersRequest, context: grpc.ServicerContext
+    ) -> get_user_proto.UsersResponse:
         """
         Get all existing users
 
         Parameters
         ----------
-        _
-            Empty request
-        context
+        request : get_user_proto.GetAllUsersRequest
+            Requesting user's data
+        context : grpc.ServicerContext
             Request context
 
         Returns
         -------
         UsersResponse
             Response object with array of user data
+
         """
         try:
+            requesting_user = User.from_grpc_user(request.requested_user)
+            if requesting_user.type != UserType.ADMIN:
+                context.set_code(grpc.StatusCode.UNAUTHENTICATED)
+                return get_user_proto.UsersResponse(
+                    status_code=403,
+                    message="Unauthorized",
+                )
             users = await self._user_repository.get_all_users()
             context.set_code(grpc.StatusCode.OK)
             return get_user_proto.UsersResponse(
@@ -387,7 +396,7 @@ class IdentityServiceImpl(GrpcServicer):
         Parameters
         ----------
         request : UpdateUserRequest
-            User data to be updated and current user id
+            User data to be updated and current user data
         context : grpc.ServicerContext
             Request context
 
@@ -399,7 +408,15 @@ class IdentityServiceImpl(GrpcServicer):
         """
         try:
             user = User.from_update_grpc_user(grpc_user=request.new_user)
+            requesting_user = User.from_grpc_user(request.requesting_user)
             db_user = await self._user_repository.get_user_by_id(user_id=user.id)
+
+            if requesting_user.type != UserType.ADMIN and requesting_user.id != user.id:
+                context.set_code(grpc.StatusCode.UNAUTHENTICATED)
+                return auth_proto.CredentialsResponse(
+                    status_code=403,
+                    message="Unauthorized",
+                )
 
             if self._encoder.compare(user.password, db_user.password):
                 user.password = db_user.password
@@ -451,7 +468,7 @@ class IdentityServiceImpl(GrpcServicer):
         Parameters
         ----------
         request : DeleteUserRequest
-            ID of user to be deleted
+            ID of user to be deleted and current user data
         context : grpc.ServicerContext
             Request context
 
@@ -462,10 +479,19 @@ class IdentityServiceImpl(GrpcServicer):
 
         """
         try:
+            requesting_user = User.from_grpc_user(request.requesting_user)
+            user = await self._user_repository.get_user_by_id(request.user_id)
+            if requesting_user.type != UserType.ADMIN and requesting_user.id != user.id:
+                context.set_code(grpc.StatusCode.UNAUTHENTICATED)
+                return requests_proto.BaseResponse(
+                    status_code=403, message="Unauthorized"
+                )
             await self._token_repository.delete_all_refresh_tokens(
                 user_id=request.user_id
             )
             await self._user_repository.delete_user(user_id=request.user_id)
+            context.set_code(grpc.StatusCode.OK)
+            return requests_proto.BaseResponse(status_code=200)
         except ValueNotFoundError:
             context.set_code(grpc.StatusCode.NOT_FOUND)
             return requests_proto.BaseResponse(
@@ -476,9 +502,6 @@ class IdentityServiceImpl(GrpcServicer):
             return requests_proto.BaseResponse(
                 status_code=500, message="Internal server error"
             )
-
-        context.set_code(grpc.StatusCode.OK)
-        return requests_proto.BaseResponse(status_code=200)
 
     async def logout(
         self, request: auth_proto.AccessToken, context: grpc.ServicerContext
@@ -518,9 +541,9 @@ class IdentityServiceImpl(GrpcServicer):
 
         Parameters
         ----------
-        session_id
+        session_id : str
             Id of the session
-        user_id
+        user_id : str
             Id of the user
 
         Returns
